+++ datatypes/src/commonMain/kotlin/at/asitplus/crypto/datatypes/asn1/Asn1Encoder.kt
@@ -207,7 +207,7 @@ fun CryptoPublicKey.encodeToTlv() = when (this) {
                         n.ensureSize(bits.number / 8u)
                             .let { if (it.first() == 0x00.toByte()) it else byteArrayOf(0x00, *it) })
                 }
-                int { e.toInt() }
+                long { e.toLong() }
             })
 
         }
